Committing python scripts used for investigating any trends in campaign finance and bill success as well as the committee co-membership network.
There are two classes for Bill and Candidate objects which can be updated and expanded to fill needs. The file neeral_analyse_graph contains code that I used for aggregating and combining data from different sources and plotting as well. The other files are mainly for I/O; some parse input files in JSON or YAML format; others take this, extract properties and create objects.

As an informational note, FEC is the source of campaign finance data.